procinfo: Do not treat username issue as an error (#34)
If the system cannot fetch username of the minio process owner, it gets
reported as an error in the health report. However, this can happen for
legitimate reasons (e.g. in minio-operator based environments) and is
not something that needs to be flagged as an error.

Fixed by ignoring error when trying to get the username,
Set username to "<non-root>" in case of error