Detect dynamic reset in `DelConfigKV
Check if the `config applied` header (`x-minio-config-applied`) is
present in the response of the delete config api, and return an
additional bool value (restart) accordingly.

This can then be used by `mc` to decide whether to show a message to the
user asking them to restart minio for the changes to take effect.