Change key names in health data type map
The names for perf related tests were changed from perfdrive to
driveperf and perfnet to netperf, but corresponding entries were not
updated in HealthDataTypesMap.

Because of this, the `--test` flag of the `mc support diag` command
continues to expect old values. (This is a hidden flag)

Fixed by updating the required entries.