Inspect to support new inspect format
With the new inspect model, the data will be encrypted with a public
key. The exisiting symmetric key is not needed anymore, but this API
will still support it.

If a public key is specified in the API, use a POST request with
multipart form data to upload all data, otherwise use the old style.

Return a nil key if the public key is provided since the caller will be
able to decrypt data with his private key.